count.c: Support Table_connector with more than 1G entries
Also:
Change table stats access variables to uint64_t (as size_t on 32 bit CPU
is too small for them) and also some other variable to the equivalent
uintX_t definition for readability.

+++ link-grammar/parse/count.h
@@ -20,7 +20,7 @@ typedef struct count_context_s count_context_t;
 
 Count_bin* table_lookup(count_context_t *, int, int,
                         const Connector *, const Connector *,
-                        unsigned int, unsigned int *);
+                        unsigned int, size_t *);
 int do_parse(Sentence, fast_matcher_t*, count_context_t*, Parse_Options);
 bool no_count(count_context_t *, int, Connector *, unsigned int, unsigned int);
